all go files moved to folder internal
