+++ internal/constants.go
@@ -1,4 +1,4 @@
-package ispend
+package internal
 
 const (
 	IPAddress   = "localhost"
