+++ internal/graphite_client.go
@@ -1,4 +1,4 @@
-package ispend
+package internal
 
 import (
 	"bytes"
