+++ internal/graphite_metric.go
@@ -1,4 +1,4 @@
-package ispend
+package internal
 
 import (
 	"fmt"
