+++ internal/login_session.go
@@ -1,4 +1,4 @@
-package ispend
+package internal
 
 type LoginSession struct {
 	Username  string
