+++ internal/spending_handler.go
@@ -1,4 +1,4 @@
-package ispend
+package internal
 
 import (
 	"net/http"
