+++ internal/user.go
@@ -1,4 +1,4 @@
-package ispend
+package internal
 
 type Users []*User
