+++ scripts/redeploy.sh
@@ -1,4 +1,4 @@
 #!/bin/bash
 curl localhost:5001/harakiri
-sleep 1s
-go run cmd/main.go -port=5001 -env=p -logfile=ispend.log &
+sleep 2s
+go run cmd/main.go -port=5001 -logfile=ispend.log &
